refactor(api): split domain-analysis snapshot readers out of cache module
The fast-path commit pushed domain-analysis-cache.ts to 730 lines, past
the 700-line hard limit. Move readModelAgreementSnapshotStateFromSnapshot
to its own snapshot-readers module — it's a read accessor used by other
resolvers, not part of the cache orchestration. Drop the dead
readCellLevelOutcomesFromSnapshot (superseded by the state reader).
